pyproject.toml
[project]
name = "arknights-avatar"
version = "0.1.0"
description = "Making Arknights operator avatars"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=12.0.0",
]
